Move sum compute logic to component folder.
